+++ MordoEngine/res/shaders/basic.fs
@@ -1,11 +1,11 @@
 #version 330 core
 
 in vec2 TexCoord;
-in vec3 vColor;
-
 out vec4 FragColor;
+
 uniform sampler2D texture1;
 
-void main() {
+void main()
+{
     FragColor = texture(texture1, TexCoord);
 }